requirements.txt
# Core ML
tensorflow>=2.13.0
keras>=2.13.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
opencv-python>=4.8.0
Pillow>=10.0.0
pandas>=2.0.0

# API
fastapi>=0.103.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
